requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.24
pydantic>=2.5
pydantic-settings>=2.1
requests>=2.31
numpy>=1.24
httpx>=0.24
pytest>=7.4
sentence-transformers>=2.2.2
openai>=1.40.0


